Make GTMOAuth use ARC.
Depend on Tidbits, so that we can drop the NSData+Base64
implementation here and use the one in Tidbits (they're both
derived from Matt Gallagher's implementation).  In particular,
the Tidbits one uses ARC.

Use NSUUID to create the nonce, instead of CFUUIDCreate.  Use
gtm_stringByEscapingForURLArgument instead of the
CFURLCreateStringByAddingPercentEscapes-based implementation.